When the user is signed in with google button 'Connect with Google+ disapears'.
Using the API

Just like with the Facebook PHP SDK, the Google JavaScript SDK now has an access token that it's storing. This means we can start making API calls. I'll copy in a function that uses the API to get a list of all of the people in my circles and print their smiling faces.

This looks for a div with the id google-plus-farmers, so let's add that to our page.

Ok, let's try it! When we refresh and sign in, we get a beautiful box of farmers in our circle! In my console, if we click on the AJAX call that was made, we can see that an access token was sent on the Authorization: Bearer header. OAuth is happening behind the scenes!