Including the JavaScript SDK The implicit OAuth flow can be done without any tools, but Google makes our life a lot easier by giving us a JavaScript SDK(https://developers.google.com/+/web/signin/javascript-flow#step_2_include_the_google_script_on_your_page).
This exposes a global gapi object we'll use in a second.

Initiate the Sign-in Flow
Let's add a 'Connect with Google+' button on the homepage and attach a jQuery click event listener to it.

Loads the client library and the auth2 library together for efficiency. Loading the auth2 library is optional here since  function will load it if not already loaded. Loading it upfront can save one network request.

Initialize the client with API key and Google+ API and initialize OAuth with an OAuth 2.0 client ID and scopes (space delimited string) to request access. You have a bunch of examples at https://github.com/google/google-api-javascript-client/tree/master/samples and documentation at https://developers.google.com/+/web/api/javascript

Listen for sign-in state changes. Handle the initial sign-in state.