No Refresh Tokens in the Implicit Grant Type
Even if an OAuth server supports refresh tokens, you won't be given one if you use the implicit flow. To see what I mean, change the response_type parameter on our COOP authorize URL to token and add a die statement right at the top of the code that handles the redirect.
When we try the process again, COOP redirects us back with a URL that contains an access token instead of the authorization code:

http://localhost:9000/coop/oauth/handle#
    access_token=eaf215f677bea1562026df05ecca202163a6c69f
    &expires_in=86400
    &token_type=Bearer
    &scope=eggs-count+profile
Since this is how the implicit flow works, this no surprise. But notice that there's no refresh token. That's one major disadvantage of using the implicit grant type.