I've already started the little script for the CRON job, which you can see at data/refresh_tokens.php. What we want to do here is use the COOP API to count and save each user's daily eggs.
But first, we need to make sure that everyone has a non-expired access token. Let's use a method called getExpiringTokens() that I've already prepared. This queries the database and returns details for all users whose coopAccessExpiresAt value is today or earlier.

 Tip
In the background, this is just running a query similar to this:

SELECT * FROM users WHERE coopAccessExpiresAt < '2014-XX-YY';
Next, let's iterate over each expiring token. To get a refresh token, we'll make an API request to the very-familiar /token endpoint. In fact, I'll start by copying the Guzzle API call from CoopOAuthController:

/ data/refresh_tokens.php
// ...

{dollar_sign}expiringTokens = {dollar_sign}conn->getExpiringTokens();

foreach ({dollar_sign}expiringTokens as {dollar_sign}userInfo) {

  {dollar_sign}request = {dollar_sign}http->post('/token', null, array(
    'client_id'     => 'TopCluck',
    'client_secret' => '2e2dfd645da38940b1ff694733cc6be6',
    'grant_type'    => 'authorization_code',
    'code'          => ,
    'redirect_uri'  => {dollar_sign}this->generateUrl('coop_authorize_redirect', array(), true),
  ));

  // make a request to the token url
  {dollar_sign}response = {dollar_sign}request->send();
  {dollar_sign}responseBody = {dollar_sign}response->getBody(true);
  {dollar_sign}responseArr = json_decode({dollar_sign}responseBody, true);

}

Of course, we don't have a {dollar_sign}code variable, but we do have the user's refresh token. Change grant_type to be refresh_token and replace the code parameter with the refresh_token. We can also remove the redirect_uri, which isn't needed with this grant type.
Let's try out the API call! Tweak the getExpiringTokens() method temporarily. We don't actually have any users with expiring tokens, but this change will return any tokens expiring in the next month, which should be everyone.

Now, try it by executing the script from the command line:

$ php data/refresh_token.php
With any luck, we should see a familiar-looking JSON response:

{
  'access_token': '1729a2fc9e6d6da2d2cb877c5bf3239fd2c57d0d',
  'expires_in': 86400,
  'token_type': 'Bearer',
  'scope': 'eggs-count profile',
  'refresh_token': 'f6ecef2bf0d16d7c13a983616b30d72ca915ab65'
}