Perfect! Now we just need to update the user with the new coopAccessToken, coopExpiresAt and coopRefreshToken. Again, we can copy or re-use some code from CoopOAuthController, since this is the same response from there. The saveNewTokens() method is a shortcut to update the user record with this data.
Tip
In the background, this is just running an UPDATE query against this user to update the access token, expiration and refresh token columns.

Let's add a little message so we can see what's going on.

But when we try it now, the script blows up! Since we're still dumping the raw response, above the exception we can see the message 'Invalid refresh token'. The problem is that when we used the refresh token a second ago, the COOP API gave us a new one and invalidated the old one. We weren't saving it yet, so now we're stuck and need to re-authorize the user.

 Tip
An OAuth server may or may not invalidate the refresh token after using it - that's totally up to the server.

Go back to the site, log out, and log back in with COOP. This will get a new refresh token for the user. And since we're saving the new refresh token, in our script each time, we can run it over and over again without any issues.

And now that we've refreshed everyone's access tokens, we could loop through each user and send an API request to count their eggs. The code for that would look almost exactly like code in the CountEggs.php file, so we'll leave that to you.

Nothing lasts Forever

Of course, nothing lasts forever, and even the refresh token will eventually expire. These tokens commonly last for 14-60 days, and afterwards, you have no choice but to ask the user to re-authorize your application.

 Tip
A refresh token could last forever - it's up to the OAuth server. However, it's still possible that the user revokes access in the future.

This means that unless your OAuth server has some sort of key that lasts forever, our CRON job will eventually not be able to count the eggs for all of our farmers. We may need to send them an email to re-authorize or be ok that these inactive users aren't updated anymore.