Since TopCluck is handling a lot of access tokens for Brent's farmer friends, he wants to make sure it's secure. Nothing would be worse than for the access tokens of the TopCluck farmers to get stolen - allowing some city slicker to take control of the good people's farms!
Exchanging tokens in a secure way isn't easy because there are a lot of opportunities for a hacker to be listening to the requests or doing some other clever thing.

CSRF Protection with the state Parameter

Open up CoopOAuthController so that we can squash a really common attack. In the authorize redirect URL, add a state parameter and set its value to something that's only known to the session for this user. We can do that by generating a random string and storing it in the session.
Let's also add a die statement in the receiveAuthorizationCode() function that's executed after COOP redirects back to us.
Log out and click to login via COOP. Of course, when we redirect to COOP, the new state parameter is there. Interestingly, after we authorize, COOP redirects back to us and also includes that exact state parameter.