In receiveAuthorizationCode(), we just need to make sure that state matches the string that we set in the session exactly. If it doesn't, let's render an error page: this could be an attack.
Using the state parameter is just like using a CSRF token with a form: it prevents XSS attacks.

When we log in now, it all still works perfectly.

Imagine I start the authorization process, but use a browser plugin to prevent COOP from redirecting me back to TopCluck. Then, I post the redirect URL with my valid authorization code to a forum somewhere, maybe embedded in an image tag. Assuming you're logged into TopCluck, when you view this page, the image tag will make a request to TopCluck, which exchanges the authorization code for an access token in the background.

So what? Well, CoopOAuthController would end up saving your coopUserId to the attacker's TopCluck account. This means when the attacker logs into TopCluck using COOP, they'll be logged in as you!

So, always use a state parameter. Fortunately, when you work with something like Facebook's SDK, this happens automatically. We didn't realize it, but it was generating a state parameter, saving it to the session, and checking it when we exchanged the authorization code for the access token. That's pretty nice.