Registering the Redirect URI
Head over to COOP and check out our application there. One field we left blank was the Redirect URI. Let's fill it in now with a made-up URL.

Try logging in again. This time, we immediately get an error from COOP:

The redirect URI provided is missing or does not match
The redirect URI is a security measure that guarantees that nobody can use your client ID, which is public, to authorize users and redirect with the authorization code or access token back to their site. Many OAuth servers require this to be filled in. In fact, we saw that with Facebook earlier

I'll re-edit the application and put in our exact redirect_uri value. When we try to login in now, it works.

Most OAuth servers will require this value. Sometimes, the URL we put here must match the redirect_uri parameter exactly. Other times, it's a fuzzy match. This is up to the OAuth server you're using, but exact matching is much more difficult to fake.

In a client-side environment where the code or token is passed via JavaScript, the OAuth server may just ask you for your hostname or a list of JavaScript origins. These function the same way: to prevent JavaScript on some other hostname from using your client id.