The Insecurity of Implicit
The implicit grant type is the least secure grant type because the access token can be read by other JavaScript on your page and could be a victim of XSS attacks. If you decide to use implicit, you must be extra careful in preventing the attacks on the pages where access tokens are used in JavaScript.

This is another example of why registering an exact redirect URI is important. If an attacker locates just one XSS vulnerability on your site, they could manipulate the redirect URI to point there, and use it to steal access tokens. It's also even more important to validate your state parameter.

If it's at all possible to use the authorization code grant type instead, this is much better because even if there was a man in the middle or piece of JavaScript reading your authorization code, the client secret is still needed to turn that into an access token.

One interesting thing about the implicit grant type is that the access token is passed back as a URL fragment instead of a query parameter

We didn't see this with Google+ because it was all being handled in the background for us. But this is really important because anything after the hash in a URL isn't actually sent when your browser requests a page. The JavaScript on your page can read this, but since it's not sent over the web, anyone listening between the user and the server won't be able to intercept it. That's not as important with the code, because the man-in-the-middle would still need the client secret to do anything with it.